\name{branch_length}

\alias{branch_length}

  
\title{branch_length}

\description{
get the branch length
}

\usage{
branch_length(tree, samp, adjust, adjfun=log10 )
}

\arguments{
\item{tree}{??}
\item{samp}{??}
\item{adjust}{logical. If \code{TRUE} then ...}
\item{adjfun}{function.}
}

\value{
numeric 
}

\examples{
2 + 2
#branch_length( ) 
}

\keyword{spatial}

