\name{get_samplesize}

\alias{get_samplesize}

  
\title{get_samplesize}

\description{
get the branch length
}

\usage{
get_samplesize(range, fun=n_zones, ...)
}

\arguments{
\item{range}{SpatVector or SpatRaster}
\item{fun}{function that computes the number of zones. The default is \code{\link{n_zones}}}
\item{...}{additional arguments for \code{fun}}
}

\value{
numeric 
}

\examples{
2 + 2
#get_samplesize( ) 
}

\keyword{spatial}

