\name{n_zones}

\alias{n_zones}

  
\title{n_zones}

\description{
get the branch length
}

\usage{
n_zones(x, min_area, m=3)
}

\arguments{
\item{x}{numeric, ranze sizes}
\item{min_area}{numeric. Minimal size of a range}
\item{m}{positive numeric. Multiplier}
}

\value{
numeric 
}

\examples{
2 + 2
#n_zones( ) 
}

\keyword{spatial}

